Refactor the terrafom-deployment && fix some comments.
