+++ modules/database/src/main/java/org/eclipse/xpanse/modules/database/service/DeployResourceRepository.java
@@ -6,7 +6,6 @@
 
 package org.eclipse.xpanse.modules.database.service;
 
-import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -14,6 +13,6 @@ import org.springframework.stereotype.Repository;
  * Interface to access default JPA methods.
  */
 @Repository
-public interface DeployResourceRepository extends JpaRepository<DeployResourceEntity, UUID> {
+public interface DeployResourceRepository extends JpaRepository<DeployResourceEntity, String> {
 
 }
