+++ modules/deployment/src/main/java/org/eclipse/xpanse/modules/deployment/deployers/terraform/TerraformProviders.java
@@ -4,12 +4,12 @@
  *
  */
 
-package org.eclipse.xpanse.modules.terraform.deployment;
+package org.eclipse.xpanse.modules.deployment.deployers.terraform;
 
 import java.util.Objects;
+import org.eclipse.xpanse.modules.deployment.deployers.terraform.providers.Aws;
+import org.eclipse.xpanse.modules.deployment.deployers.terraform.providers.Huawei;
 import org.eclipse.xpanse.modules.ocl.loader.data.models.enums.Csp;
-import org.eclipse.xpanse.modules.terraform.deployment.providers.Aws;
-import org.eclipse.xpanse.modules.terraform.deployment.providers.Huawei;
 
 /**
  * Terraform providers.
