+++ modules/deployment/src/main/java/org/eclipse/xpanse/modules/deployment/deployers/terraform/providers/Aws.java
@@ -4,22 +4,23 @@
  *
  */
 
-package org.eclipse.xpanse.modules.terraform.deployment.providers;
+package org.eclipse.xpanse.modules.deployment.deployers.terraform.providers;
 
+
+import org.eclipse.xpanse.modules.deployment.deployers.terraform.Provider;
 import org.eclipse.xpanse.modules.ocl.loader.data.models.enums.Csp;
-import org.eclipse.xpanse.modules.terraform.deployment.Provider;
 
 /**
- * Terraform provider information for Huawei.
+ * Terraform provider information for AWS.
  */
-public class Huawei implements Provider {
+public class Aws implements Provider {
 
     public static final String PROVIDER = """
             terraform {
               required_providers {
-                huaweicloud = {
-                  source = "huaweicloud/huaweicloud"
-                  version = ">= 1.20.0"
+                aws = {
+                  source  = "hashicorp/aws"
+                  version = "~> 4.0"
                 }
               }
             }
@@ -32,7 +33,9 @@ public class Huawei implements Provider {
 
     @Override
     public Csp getCsp() {
-        return Csp.HUAWEI;
+        return Csp.AWS;
     }
 
 }
+
+
