+++ modules/deployment/src/main/java/org/eclipse/xpanse/modules/deployment/deployers/terraform/resource/TfStateResourceInstance.java
@@ -4,7 +4,7 @@
  *
  */
 
-package org.eclipse.xpanse.modules.terraform.deployment.resource;
+package org.eclipse.xpanse.modules.deployment.deployers.terraform.resource;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.Map;
